Added bots folder and moved webhooks
